+++ HISTORY.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+unreleased
+----------
+* Move pytogo Go code into portforward
+
 0.4.4 (2023-02-28)
 ------------------
 * Fix endless waiting